# backend/src/main/java/com/tuhungkien/ecommerce/dto/review/ReviewResponse.java
package com.tuhungkien.ecommerce.dto.review;

import lombok.Data;

import java.time.LocalDate;

@Data
public class ReviewResponse {
    private Long id;
    private String author;
    private String message;
    private Integer rating;
    private LocalDate date;
}